legal_document_analyzer/requirements.txt
streamlit==1.28.1
transformers==4.35.2
torch==2.1.1
PyMuPDF==1.23.8
python-docx==0.8.11
spacy==3.7.2
pandas==2.1.3
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
requests==2.31.0
Pillow==10.1.0
plotly==5.17.0
nltk==3.8.1
scikit-learn==1.3.2